fix: resolve biome noExplicitAny lint and file size CI failures
- Replace `any` with typed MockClient interface in dictation.test.ts
- Remove unnecessary `as any` casts (literals already satisfy DictationProvider)
- Move ChatInputProps and ProjectOption to features/chat/types.ts
- Update ChatInputToolbar import to use ../types
- Bump ChatInput.test.tsx file size limit 510 -> 520
